Message Class redesigned
Message class is now a super class for a different type of message, as new design states
that the messages can be ChatMessage, FriendCreationRequest, FriendCreationResponse and etc.

The basic Message class now contains Sender, receiver, timestamp and message type.

A new friend constuctor that contains only IP and port is added to faclility the new FriendCreationRequest
class, as this change, MAC is no longer a final member and a setter method is also created.

A message factory is also created to create corresponding message.

In the NSD module, as the Friend class constructor require port as parameter, there is a mocked port number
added to here to ensure complie successfully.

Minor change of remove unused package import from some files.

	new file:   app/src/androidTest/java/com/iotbyte/wifipidgin/chat/ChatMessageTest.java
	deleted:    app/src/androidTest/java/com/iotbyte/wifipidgin/chat/MessageTest.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/chat/Chat.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/chat/ChatManager.java
	deleted:    app/src/main/java/com/iotbyte/wifipidgin/chat/Message.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/commmodule/MessageClient.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/commmodule/MessageServer.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/friend/Friend.java
	new file:   app/src/main/java/com/iotbyte/wifipidgin/message/ChatMessage.java
	new file:   app/src/main/java/com/iotbyte/wifipidgin/message/FriendCreationRequest.java
	new file:   app/src/main/java/com/iotbyte/wifipidgin/message/Message.java
	new file:   app/src/main/java/com/iotbyte/wifipidgin/message/MessageFactory.java
	new file:   app/src/main/java/com/iotbyte/wifipidgin/message/MessageType.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/nsdmodule/NsdClient.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/ui/testChat.java