Inital commit for IncomingMessageHandling mechanism
Adding IncomingMessageHandlingService and InomingMessageHandler which constantly monitoring
the IncomingMessageQueue in ChatMananger class. The only work that the incomingMessageHandler
does is to have a running thread to monitor the queue status, as long as the queue is not empty,
it will call the dequeue method in ChatManager class. The actual handling code is in ChatManager's
dequeue method.

 Changes to be committed:
	new file:   app/src/androidTest/java/com/iotbyte/wifipidgin/chat/IncomingMessageHandlingServiceTest.java
	modified:   app/src/main/AndroidManifest.xml
	modified:   app/src/main/java/com/iotbyte/wifipidgin/chat/ChatManager.java
	new file:   app/src/main/java/com/iotbyte/wifipidgin/chat/IncomingMessageHandler.java
	new file:   app/src/main/java/com/iotbyte/wifipidgin/chat/IncomingMessageHandlingService.java
	new file:   app/src/main/java/com/iotbyte/wifipidgin/chat/IncomingMessageQueueListener.java
	modified:   app/src/main/res/values/strings.xml