Event registration with ChannelDao DaoEventBoard
ChannelManager now listen to ChannelDao's event change, which will update the ChannelManager's list of
Channel.

A new ChannelDatabaseChangeListener is set in the event of ChannelDao's data change from ChannelManager.
The intension of this listener is provide Channel UI a way to kick in UI refresh when some change happens to the
channels on the backend that does not come from the UI.

 Changes to be committed:
        new file:   app/src/main/java/com/iotbyte/wifipidgin/channel/ChannelDatabaseChangeListener.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/channel/ChannelManager.java