Fixes 0000007
Code changed. Update user self's MAC everytime message server luanches. BUT, need to discuss with DQ about how to get "mySelf" from DB. Currently we are already Friend which we cannot cast to mySelf to use the setMac() method.

fix a bug in Util for converting byte array to string for MAC.