Fixed#0000010, ChannelCreationMessage feature
ChannelCreationMessage feature complete:

The ChannelCreationMessage class complete.
A sendChannelCreationMessageToAll() method is added to ChannelMananger to make sending of this message easier
The handling fo ChannelCreationMessage is added into ChatManager.
However,a UI/notification is needed to make user to select the join/decline of the channelcreation
Also, the channel blocking feature need to be adding here as well.

some TODO and FIXME are added here to make it easy

A mocked handling is added here which assue the user always choose YES

 Changes to be committed:
	modified:   app/src/main/java/com/iotbyte/wifipidgin/channel/ChannelManager.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/chat/ChatManager.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/message/ChannelCreationMessage.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/message/MessageFactory.java