Fixed#0000020 ChannelCreationRequest/Response
The channel Creation process will be changed as of this new design.

A creator of a channel will send a ChannelCreationRequest to all people he/she choose to add in the channel. This list of all receiver of the ChannelCreationRequest is only temporary as some of the people may not accept this invite

A new channel will be created here as he send out this request, however, the only friend in this channel is only the creator himself.

when a friend receive this invitation, he can choose to response or not, if he response, that means he would be join the channel. a channelCreationResponse will be created and send back to the creator of the channel.

Upon the creator receive the channelCreationResponse, he will update the friend list in the channel and send a ChannelCreationRequest along with the updated friends lists to all friends in his local channel.

NOTE in here the ChannelCreationRequest is used to send request and update the information of the channel as well. It is a dual functional Requests

When a friend receive this message, he will update his local channel friend list.

	modified:   app/src/main/java/com/iotbyte/wifipidgin/channel/Channel.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/channel/ChannelManager.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/chat/ChatManager.java
	renamed:    app/src/main/java/com/iotbyte/wifipidgin/message/ChannelCreationMessage.java -> app/src/main/java/com/iotbyte/wifipidgin/message/ChannelCreationRequest.java
	new file:   app/src/main/java/com/iotbyte/wifipidgin/message/ChannelCreationResponse.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/message/Message.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/message/MessageFactory.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/message/MessageType.java