Fix issue in Saving Friend&channel in ChannelMgr
A bug is found in saveAChannelToDataBase and saveChannelInfoToDataBase

Where possible that when a friend or channel has no assigned ID, it possible
means the friend/channel are never exist in DB, so just add it.
But it also possible means the channel/friend already exist in DB, but the
current friend/channel was passed from a creationRequest, then need to find the
ID from DB then do a update.

Another issue is found in ChatManager when handle CHANNEL_CREATION_REQUEST, a return value
is missing if the channel already exist.

if this is missing, the switch statment will continue to the next case, and yield a cast issue

A update to gradle version is also commited.

 Changes to be committed:
	modified:   app/src/main/java/com/iotbyte/wifipidgin/channel/ChannelManager.java
	modified:   app/src/main/java/com/iotbyte/wifipidgin/chat/ChatManager.java
	modified:   build.gradle